Added small documentation, fixed bug in GoBoard::countLiberties,
countLiberties was returning too early causing only part of a group to be
searched and marked so that a later search of group would be tricked that
it was dead. Moved the player enum type to Stone.h to allow it to be used
universally rather that just in GoBoard.