Made board traversal methods easier to read/understand, code is effectively the same
Changed the board traversal functions to use an array of x,y coordinate pairs
and a loop to shorten the functions down to single block from 4, the code
executes exactly as before, but is easier to read and edit in the future.
Area scoring still needs thorough testing but appears to be working correctly.